changed default coeff type to Int

+++ src/dirac/diracoperators.jl
@@ -49,8 +49,8 @@ import Base: show,
     getket(op::DiracOperator) = op.ket
     getbra(op::DiracOperator) = op.bra
 
-    coefftype(op::DiracOperator) = Complex128
-    coefftype{O<:DiracOperator}(::Type{O}) = Complex128
+    coefftype(op::DiracOperator) = Int
+    coefftype{O<:DiracOperator}(::Type{O}) = Int
     
     coeff(op::DiracOperator) = one(coefftype(op))
     operator(op::DiracOperator) = op
